+++ src/main/java/com/github/unchama/seichiassist/commands/contributeCommand.java
@@ -47,7 +47,7 @@ public class contributeCommand implements TabExecutor {
 		}else if(args[0].equalsIgnoreCase("add") && args.length == 3){
 
 			//sqlをusernameで操作
-			if (databaseGateway.setContribute(sender, args[1], TypeConverter.toInt(args[2]))) {
+			if (databaseGateway.playerDataManipulator.setContribute(sender, args[1], TypeConverter.toInt(args[2]))) {
 				sender.sendMessage(ChatColor.GREEN + args[1] + "に貢献度ポイント" + args[2] + "を追加しました");
 
 				//指定プレイヤーがオンラインの場合即時反映
@@ -67,7 +67,7 @@ public class contributeCommand implements TabExecutor {
 		}else if(args[0].equalsIgnoreCase("remove") && args.length == 3){
 
 			//sqlをusernameで操作
-			if (databaseGateway.setContribute(sender, args[1], (-1 * TypeConverter.toInt(args[2])))) {
+			if (databaseGateway.playerDataManipulator.setContribute(sender, args[1], (-1 * TypeConverter.toInt(args[2])))) {
 				sender.sendMessage(ChatColor.GREEN + args[1] + "の貢献度ポイントを" + args[2] + "減少させました");
 
 				//指定プレイヤーがオンラインの場合即時反映
